Move sweeper FirebaseAndroidApp to the service package (#8186)
* Move sweeper FirebaseAndroidApp to the service package

This reverts commit aad16c50e8589dda9162acc72d44c05488a50826.

* Append the current sweeper file name to the sweeper name

* Create a utility function AddTestSweepers

* Hash the file name
