+++ google-beta/gcp_sweeper_test.go
@@ -6,6 +6,8 @@ import (
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+
+	_ "github.com/hashicorp/terraform-provider-google-beta/google-beta/services/firebase"
 )
 
 func TestMain(m *testing.M) {
