+++ google-beta/services/firebase/resource_firebase_android_app_sweeper.go
@@ -1,6 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
-package google
+package firebase
 
 import (
 	"context"
@@ -9,17 +9,13 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
 	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
 func init() {
-	resource.AddTestSweepers("FirebaseAndroidApp", &resource.Sweeper{
-		Name: "FirebaseAndroidApp",
-		F:    testSweepFirebaseAndroidApp,
-	})
+	acctest.AddTestSweepers("FirebaseAndroidApp", testSweepFirebaseAndroidApp)
 }
 
 // At the time of writing, the CI only passes us-central1 as the region
